deps: support pandas 2.2 (#492)
* deps: support pandas 2.2

* fix tests and loading multiindex

* fix doctests

* avoid circular imports

* missing import

* allow pandas in prerelease

* avoid table.execute in ibis tests

* update notebook to use native plotting

+++ packages/bigframes/tests/system/small/test_groupby.py
@@ -228,7 +228,9 @@ def test_dataframe_groupby_multi_sum(
         (lambda x: x.cumsum(numeric_only=True)),
         (lambda x: x.cummax(numeric_only=True)),
         (lambda x: x.cummin(numeric_only=True)),
-        (lambda x: x.cumprod()),
+        # pandas 2.2 uses floating point for cumulative product even for
+        # integer inputs.
+        (lambda x: x.cumprod().astype("Float64")),
         (lambda x: x.shift(periods=2)),
     ],
     ids=[
